Add a work arround for copying DBLC.exe from VOS to host
On Windows 11 22H2 we observed that copied DBLC.exe from VOS to host is
broken so that DBLC.exe related features doesn't work correctly.
It seems that it occurs on the following condition:

* Windows 11 22H2
  * It doesn't reproduce on Windows 11 21H2 or former
* Use `CopyFile()` of Win32
* Copy a file from VOS to a host's folder which is
  `DirectoryIsolationMode=Merged`

We observed such issue only on copying DBLC.exe in the current code, so
we add the work arround only for it for now.