Use win32 default dialog to select client certificate
In the previous versions, we use custom dialog
which was implemented in DlgCertification class.

It works well, but there are some non user-friendly
points to distinct client certificate
because it adopts custom user-interface which is
different from Edge or Chrome.

In this commit, we use Crypt UI API to show default
dialog. this change improves user experience.

NOTE: For backward compatibility, use custom dialog for Windows 10
because there is a bug that new tab can't be opened after accessing certstore.
See https://github.com/ThinBridge/Chronos/pull/141

Signed-off-by: Kentaro Hayashi <hayashi@clear-code.com>