Disable first party sets
About first party sets: https://privacysandbox.com/open-web/#the-privacy-sandbox-timeline/

CEF119 has a bug that CEF crashes when a mouse cursor enters to a search box of YouTube when first party sets is enabled.

https://github.com/chromiumembedded/cef/issues/3643

We diasble first party sets as temporary measure.
The bug is fixed since CEF122, so we can enable first party sets if we adopt CEF122+.