Remove EnableDeleteCache
In SG mode (ThinApp mode), a data folder is always removed on shutting down, so `EnableDeleteCache = false` doesn't work as expected,
which is a specification (limitation) of ThinApp.
Also, since CEF120, we should specify unique cache folder for each instances, that make re-using cache folder difficult.

https://magpcss.org/ceforum/viewtopic.php?f=6&t=19677&p=54689&hilit=OnAlreadyRunningAppRelaunch#p54689

Because of the reason above,  we should remove EnableDeleteCache.