Fix a bug that ChFiler and SpC does not load a specified user config file
Chronos starts SpC.exe and ChFiler with command lines like below.

```
SpC.exe "B:\" /ChronosConfig=/path/to/config/file
```

In that case, SpC gets a single argument `B:" /ChronosConfig=/path/to/config/file`
because `\"` is escaped to `"` and the first `"` is ignored as there is no corresponding
`"`, but it is not intentional. Chronos wants to pass `B:\` and `/ChronosConfig=/path/to/config/file`
to SpC.exe. To avoid that behavior, this patch removes trailing `\`s when we specify folder path.