Fix zoom size applied to CEF
The zoom size applied to CEF was not corresponding the user specified zoom scale value.
In Chrome runtime style, a current zoom size is popup on screen. As a result, the popup
zoom size mismatched to the use specified zoom scale value.

Also the implementations about zoom function are refactored.